requirements.txt
opencv-python
depthai>=2.19.1.0
depthai-sdk>=1.9.1.1
numpy>=1.18.5
oss2
Rpi.GPIO
python-dotenv